Version 2015.8.11

Conflicts:
	salt/states/pip_state.py
	salt/states/service.py

+++ salt/modules/saltutil.py
@@ -26,7 +26,7 @@ try:
 except ImportError:
     HAS_ESKY = False
 # pylint: disable=no-name-in-module
-from salt.ext.six import string_types
+import salt.ext.six as six
 from salt.ext.six.moves.urllib.error import URLError
 # pylint: enable=import-error,no-name-in-module
 
@@ -51,8 +51,15 @@ import salt.utils.event
 import salt.utils.url
 import salt.transport
 import salt.wheel
+
+HAS_PSUTIL = True
+try:
+    import salt.utils.psutil_compat
+except ImportError:
+    HAS_PSUTIL = False
+
 from salt.exceptions import (
-    SaltReqTimeoutError, SaltRenderError, CommandExecutionError
+    SaltReqTimeoutError, SaltRenderError, CommandExecutionError, SaltInvocationError
 )
 
 __proxyenabled__ = ['*']
@@ -88,7 +95,7 @@ def _sync(form, saltenv=None):
     '''
     if saltenv is None:
         saltenv = _get_top_file_envs()
-    if isinstance(saltenv, string_types):
+    if isinstance(saltenv, six.string_types):
         saltenv = saltenv.split(',')
     ret = []
     remote = set()
@@ -258,18 +265,22 @@ def update(version=None):
 
 def sync_beacons(saltenv=None, refresh=True):
     '''
-    Sync the beacons from the _beacons directory on the salt master file
-    server. This function is environment aware, pass the desired environment
-    to grab the contents of the _beacons directory, base is the default
-    environment.
-
     .. versionadded:: 2015.5.1
+
+    Sync the beacons from the ``salt://_beacons`` directory on the Salt
+    fileserver. This function is environment-aware, pass the desired
+    environment to grab the contents of the ``_beacons`` directory from that
+    environment. The default environment, if none is specified,  is ``base``.
+
+    refresh : True
+        Also refresh the beacons available to the minion.
 
     CLI Example:
 
     .. code-block:: bash
 
         salt '*' saltutil.sync_beacons
+        salt '*' saltutil.sync_beacons saltenv=dev
     '''
     ret = _sync('beacons', saltenv)
     if refresh:
@@ -279,18 +290,23 @@ def sync_beacons(saltenv=None, refresh=True):
 
 def sync_sdb(saltenv=None, refresh=False):
     '''
-    Sync sdb modules from the _sdb directory on the salt master file
-    server. This function is environment aware, pass the desired environment
-    to grab the contents of the _sdb directory, base is the default
-    environment.
-
     .. versionadded:: 2015.5.7
+
+    Sync sdb modules from the ``salt://_sdb`` directory on the Salt
+    fileserver. This function is environment-aware, pass the desired
+    environment to grab the contents of the ``_sdb`` directory from that
+    environment. The default environment, if none is specified,  is ``base``.
+
+    refresh : False
+        This argument has no affect and is included for consistency with the
+        other sync functions.
 
     CLI Example:
 
     .. code-block:: bash
 
         salt '*' saltutil.sync_sdb
+        salt '*' saltutil.sync_sdb saltenv=dev
     '''
     ret = _sync('sdb', saltenv)
     return ret
@@ -298,10 +314,13 @@ def sync_sdb(saltenv=None, refresh=False):
 
 def sync_modules(saltenv=None, refresh=True):
     '''
-    Sync the modules from the _modules directory on the salt master file
-    server. This function is environment aware, pass the desired environment
-    to grab the contents of the _modules directory, base is the default
-    environment.
+    Sync the modules from the ``salt://_modules`` directory on the Salt
+    fileserver. This function is environment-aware, pass the desired
+    environment to grab the contents of the ``_modules`` directory from that
+    environment. The default environment, if none is specified,  is ``base``.
+
+    refresh : True
+        Also refresh the execution modules available to the minion.
 
     .. important::
 
@@ -320,13 +339,12 @@ def sync_modules(saltenv=None, refresh=True):
         See :ref:`here <reloading-modules>` for a more detailed explanation of
         why this is necessary.
 
-    .. versionadded:: 2015.5.1
-
     CLI Example:
 
     .. code-block:: bash
 
         salt '*' saltutil.sync_modules
+        salt '*' saltutil.sync_modules saltenv=dev
     '''
     ret = _sync('modules', saltenv)
     if refresh:
@@ -336,16 +354,20 @@ def sync_modules(saltenv=None, refresh=True):
 
 def sync_states(saltenv=None, refresh=True):
     '''
-    Sync the states from the _states directory on the salt master file
-    server. This function is environment aware, pass the desired environment
-    to grab the contents of the _states directory, base is the default
-    environment.
+    Sync the states from the ``salt://_states`` directory on the Salt
+    fileserver. This function is environment-aware, pass the desired
+    environment to grab the contents of the ``_states`` directory from that
+    environment. The default environment, if none is specified,  is ``base``.
+
+    refresh : True
+        Also refresh the execution modules available to the minion.
 
     CLI Example:
 
     .. code-block:: bash
 
         salt '*' saltutil.sync_states
+        salt '*' saltutil.sync_states saltenv=dev
     '''
     ret = _sync('states', saltenv)
     if refresh:
@@ -355,16 +377,21 @@ def sync_states(saltenv=None, refresh=True):
 
 def sync_grains(saltenv=None, refresh=True):
     '''
-    Sync the grains from the _grains directory on the salt master file
-    server. This function is environment aware, pass the desired environment
-    to grab the contents of the _grains directory, base is the default
-    environment.
+    Sync the grains from the ``salt://_grains`` directory on the Salt
+    fileserver. This function is environment-aware, pass the desired
+    environment to grab the contents of the ``_grains`` directory from that
+    environment. The default environment, if none is specified,  is ``base``.
+
+    refresh : True
+        Also refresh the execution modules available to the minion, and refresh
+        pillar data.
 
     CLI Example:
 
     .. code-block:: bash
 
         salt '*' saltutil.sync_grains
+        salt '*' saltutil.sync_grains saltenv=dev
     '''
     ret = _sync('grains', saltenv)
     if refresh:
@@ -375,16 +402,20 @@ def sync_grains(saltenv=None, refresh=True):
 
 def sync_renderers(saltenv=None, refresh=True):
     '''
-    Sync the renderers from the _renderers directory on the salt master file
-    server. This function is environment aware, pass the desired environment
-    to grab the contents of the _renderers directory, base is the default
-    environment.
+    Sync the renderers from the ``salt://_renderers`` directory on the Salt
+    fileserver. This function is environment-aware, pass the desired
+    environment to grab the contents of the ``_renderers`` directory from that
+    environment. The default environment, if none is specified,  is ``base``.
+
+    refresh : True
+        Also refresh the execution modules available to the minion.
 
     CLI Example:
 
     .. code-block:: bash
 
         salt '*' saltutil.sync_renderers
+        salt '*' saltutil.sync_renderers saltenv=dev
     '''
     ret = _sync('renderers', saltenv)
     if refresh:
@@ -394,16 +425,20 @@ def sync_renderers(saltenv=None, refresh=True):
 
 def sync_returners(saltenv=None, refresh=True):
     '''
-    Sync the returners from the _returners directory on the salt master file
-    server. This function is environment aware, pass the desired environment
-    to grab the contents of the _returners directory, base is the default
-    environment.
+    Sync the returners from the ``salt://_returners`` directory on the Salt
+    fileserver. This function is environment-aware, pass the desired
+    environment to grab the contents of the ``_returners`` directory from that
+    environment. The default environment, if none is specified,  is ``base``.
+
+    refresh : True
+        Also refresh the execution modules available to the minion.
 
     CLI Example:
 
     .. code-block:: bash
 
         salt '*' saltutil.sync_returners
+        salt '*' saltutil.sync_returners saltenv=dev
     '''
     ret = _sync('returners', saltenv)
     if refresh:
@@ -413,16 +448,19 @@ def sync_returners(saltenv=None, refresh=True):
 
 def sync_proxymodules(saltenv=None, refresh=False):
     '''
-    Sync the proxy modules from the _proxy directory on the salt master file
-    server. This function is environment aware, pass the desired environment
-    to grab the contents of the _returners directory, base is the default
-    environment.
+    .. versionadded:: 2015.8.2
+
+    Sync the proxy modules from the ``salt://_proxy`` directory on the Salt
+    fileserver. This function is environment-aware, pass the desired
+    environment to grab the contents of the ``_proxy`` directory from that
+    environment. The default environment, if none is specified,  is ``base``.
 
     CLI Example:
 
     .. code-block:: bash
 
         salt '*' saltutil.sync_proxymodules
+        salt '*' saltutil.sync_proxymodules saltenv=dev
     '''
     ret = _sync('proxy', saltenv)
     if refresh:
@@ -432,16 +470,20 @@ def sync_proxymodules(saltenv=None, refresh=False):
 
 def sync_output(saltenv=None, refresh=True):
     '''
-    Sync the output modules from the _output directory on the salt master file
-    server. This function is environment aware. Pass the desired environment
-    to grab the contents of the _output directory. Base is the default
-    environment.
+    Sync the output modules from the ``salt://_output`` directory on the Salt
+    fileserver. This function is environment-aware, pass the desired
+    environment to grab the contents of the ``_output`` directory from that
+    environment. The default environment, if none is specified, is ``base``.
+
+    refresh : True
+        Also refresh the execution modules available to the minion.
 
     CLI Example:
 
     .. code-block:: bash
 
         salt '*' saltutil.sync_output
+        salt '*' saltutil.sync_output saltenv=dev
     '''
     ret = _sync('output', saltenv)
     if refresh:
@@ -453,16 +495,22 @@ sync_outputters = salt.utils.alias_function(sync_output, 'sync_outputters')
 
 def sync_utils(saltenv=None, refresh=True):
     '''
-    Sync utility source files from the _utils directory on the salt master file
-    server. This function is environment aware, pass the desired environment
-    to grab the contents of the _utils directory, base is the default
-    environment.
+    .. versionadded:: 2014.7.0
+
+    Sync utility source files from the ``salt://_utils`` directory on the Salt
+    fileserver. This function is environment-aware, pass the desired
+    environment to grab the contents of the ``_utils`` directory from that
+    environment. The default environment, if none is specified, is ``base``.
+
+    refresh : True
+        Also refresh the execution modules available to the minion.
 
     CLI Example:
 
     .. code-block:: bash
 
         salt '*' saltutil.sync_utils
+        salt '*' saltutil.sync_utils saltenv=dev
     '''
     ret = _sync('utils', saltenv)
     if refresh:
@@ -474,16 +522,18 @@ def sync_log_handlers(saltenv=None, refresh=True):
     '''
     .. versionadded:: 2015.8.0
 
-    Sync utility source files from the _log_handlers directory on the salt master file
-    server. This function is environment aware, pass the desired environment
-    to grab the contents of the _log_handlers directory, base is the default
-    environment.
+    Sync utility source files from the ``salt://_log_handlers`` directory on
+    the Salt fileserver. This function is environment-aware, pass the desired
+    environment to grab the contents of the ``_log_handlers`` directory from
+    that environment. The default environment, if none is specified,  is
+    ``base``.
 
     CLI Example:
 
     .. code-block:: bash
 
         salt '*' saltutil.sync_log_handlers
+        salt '*' saltutil.sync_log_handlers saltenv=dev
     '''
     ret = _sync('log_handlers', saltenv)
     if refresh:
@@ -491,8 +541,47 @@ def sync_log_handlers(saltenv=None, refresh=True):
     return ret
 
 
+def sync_pillar(saltenv=None, refresh=True):
+    '''
+    .. versionadded:: 2015.8.11,2016.3.2
+
+    Sync pillar modules from the ``salt://_pillar`` directory on the Salt
+    fileserver. This function is environment-aware, pass the desired
+    environment to grab the contents of the ``_pillar`` directory from that
+    environment. The default environment, if none is specified,  is ``base``.
+
+    refresh : True
+        Also refresh the execution modules available to the minion, and refresh
+        pillar data.
+
+    .. note::
+        This function will raise an error if executed on a traditional (i.e.
+        not masterless) minion
+
+    CLI Examples:
+
+    .. code-block:: bash
+
+        salt '*' saltutil.sync_pillar
+        salt '*' saltutil.sync_pillar saltenv=dev
+    '''
+    if __opts__['file_client'] != 'local':
+        raise CommandExecutionError(
+            'Pillar modules can only be synced to masterless minions'
+        )
+    ret = _sync('pillar', saltenv)
+    if refresh:
+        refresh_modules()
+        refresh_pillar()
+    return ret
+
+
 def sync_all(saltenv=None, refresh=True):
     '''
+    .. versionchanged:: 2015.8.11,2016.3.2
+        On masterless minions, pillar modules are now synced, and refreshed
+        when ``refresh`` is set to ``True``.
+
     Sync down all of the dynamic modules from the file server for a specific
     environment. This function synchronizes custom modules, states, beacons,
     grains, returners, output modules, renderers, and utils.
@@ -522,6 +611,7 @@ def sync_all(saltenv=None, refresh=True):
     .. code-block:: bash
 
         salt '*' saltutil.sync_all
+        salt '*' saltutil.sync_all saltenv=dev
     '''
     log.debug('Syncing all')
     ret = {}
@@ -536,8 +626,12 @@ def sync_all(saltenv=None, refresh=True):
     ret['utils'] = sync_utils(saltenv, False)
     ret['log_handlers'] = sync_log_handlers(saltenv, False)
     ret['proxymodules'] = sync_proxymodules(saltenv, False)
+    if __opts__['file_client'] == 'local':
+        ret['pillar'] = sync_pillar(saltenv, False)
     if refresh:
         refresh_modules()
+        if __opts__['file_client'] == 'local':
+            refresh_pillar()
     return ret
 
 
@@ -766,11 +860,18 @@ def signal_job(jid, sig):
 
         salt '*' saltutil.signal_job <job id> 15
     '''
+    if HAS_PSUTIL is False:
+        log.warning('saltutil.signal job called, but psutil is not installed. '
+                    'Install psutil to ensure more reliable and accurate PID '
+                    'management.')
     for data in running():
         if data['jid'] == jid:
             try:
+                if HAS_PSUTIL:
+                    for proc in salt.utils.psutil_compat.Process(pid=data['pid']).children(recursive=True):
+                        proc.send_signal(sig)
                 os.kill(int(data['pid']), sig)
-                if 'child_pids' in data:
+                if HAS_PSUTIL is False and 'child_pids' in data:
                     for pid in data['child_pids']:
                         os.kill(int(pid), sig)
                 return 'Signal {0} sent to job {1} at pid {2}'.format(
@@ -989,7 +1090,7 @@ def runner(_fun, **kwargs):
     return rclient.cmd(_fun, kwarg=kwargs)
 
 
-def wheel(_fun, **kwargs):
+def wheel(_fun, *args, **kwargs):
     '''
     Execute a wheel module (this function must be run on the master)
 
@@ -997,6 +1098,13 @@ def wheel(_fun, **kwargs):
 
     name
         The name of the function to run
+
+    args
+        Any positional arguments to pass to the wheel function. A common example
+        of this would be the ``match`` arg needed for key functions.
+
+        .. versionadded:: v2015.8.11
+
     kwargs
         Any keyword arguments to pass to the wheel function
 
@@ -1004,10 +1112,33 @@ def wheel(_fun, **kwargs):
 
     .. code-block:: bash
 
-        salt '*' saltutil.wheel key.accept match=jerry
+        salt '*' saltutil.wheel key.accept jerry
     '''
-    wclient = salt.wheel.WheelClient(__opts__)
-    return wclient.cmd(_fun, kwarg=kwargs)
+    if __opts__['__role'] == 'minion':
+        master_config = os.path.join(os.path.dirname(__opts__['conf_file']),
+                                     'master')
+        master_opts = salt.config.client_config(master_config)
+        wheel_client = salt.wheel.WheelClient(master_opts)
+    else:
+        wheel_client = salt.wheel.WheelClient(__opts__)
+
+    # The WheelClient cmd needs args, kwargs, and pub_data separated out from
+    # the "normal" kwargs structure, which at this point contains __pub_x keys.
+    pub_data = {}
+    valid_kwargs = {}
+    for key, val in six.iteritems(kwargs):
+        if key.startswith('__'):
+            pub_data[key] = val
+        else:
+            valid_kwargs[key] = val
+
+    try:
+        ret = wheel_client.cmd(_fun, arg=args, pub_data=pub_data, kwarg=valid_kwargs)
+    except SaltInvocationError:
+        raise CommandExecutionError('This command can only be executed on a minion '
+                                    'that is located on the master.')
+
+    return ret
 
 
 # this is the only way I could figure out how to get the REAL file_roots
